Update the Whole Code
The whole Code is modified. 

The Coding Bugs removed.
The Code and the Algorithm is made much Simpler, and perhaps more efficient. 
Some of the Variables are renamed (according to the Need)
The main Algorithm and Technique remains the Same, just modified a little. 

The Interface is now Cleaner and more User-Friendly. 
The "How to Play" is clearly mentioned. 

It now uses X and O for Human and CPU respectively (instead of 1 and 2 which doesn't look good).
It now Decides, who Won or who Lose. 
It now Introduces the Turn, for who goes First. 
It now has a Goal (previously it was a Goal-less Game) Its Insane!